replace old streams and memory_stream
