+++ mn/include/mn/Bytes.h
@@ -22,97 +22,97 @@ namespace mn
 	}
 
 	inline static void
-	bytes_free(Bytes& self)
+	bytes_free(Bytes self)
 	{
 		memory_stream_free(self);
 	}
 
 	inline static size_t
-	bytes_size(const Bytes& self)
+	bytes_size(Bytes self)
 	{
 		return memory_stream_size(self);
 	}
 
 	inline static bool
-	bytes_eof(const Bytes& self)
+	bytes_eof(Bytes self)
 	{
 		return memory_stream_eof(self);
 	}
 
 	inline static void
-	bytes_concat(Bytes& self, const Bytes& other)
+	bytes_concat(Bytes self, Bytes other)
 	{
-		memory_stream_write(self, block_from(other.str));
+		memory_stream_write(self, block_from(other->str));
 	}
 
 	inline static void
-	bytes_rewind(Bytes& self)
+	bytes_rewind(Bytes self)
 	{
 		memory_stream_cursor_to_start(self);
 	}
 
 	inline static int64_t
-	bytes_cursor_pos(const Bytes& self)
+	bytes_cursor_pos(Bytes self)
 	{
 		return memory_stream_cursor_pos(self);
 	}
 
 	inline static void
-	bytes_cursor_move(Bytes& self, int64_t offset)
+	bytes_cursor_move(Bytes self, int64_t offset)
 	{
 		memory_stream_cursor_move(self, offset);
 	}
 
 	inline static void
-	bytes_cursor_set(Bytes& self, int64_t absolute)
+	bytes_cursor_set(Bytes self, int64_t absolute)
 	{
 		memory_stream_cursor_set(self, absolute);
 	}
 
 	inline static void
-	bytes_push8(Bytes& self, uint8_t v)
+	bytes_push8(Bytes self, uint8_t v)
 	{
 		[[maybe_unused]] size_t written_size = memory_stream_write(self, block_from(v));
 		assert(written_size == sizeof(v));
 	}
 
 	inline static void
-	bytes_push16(Bytes& self, uint16_t v)
+	bytes_push16(Bytes self, uint16_t v)
 	{
 		[[maybe_unused]] size_t written_size = memory_stream_write(self, block_from(v));
 		assert(written_size == sizeof(v));
 	}
 
 	inline static void
-	bytes_push32(Bytes& self, uint32_t v)
+	bytes_push32(Bytes self, uint32_t v)
 	{
 		[[maybe_unused]] size_t written_size = memory_stream_write(self, block_from(v));
 		assert(written_size == sizeof(v));
 	}
 
 	inline static void
-	bytes_push64(Bytes& self, uint64_t v)
+	bytes_push64(Bytes self, uint64_t v)
 	{
 		[[maybe_unused]] size_t written_size = memory_stream_write(self, block_from(v));
 		assert(written_size == sizeof(v));
 	}
 
 	inline static void
-	bytes_push32f(Bytes& self, float v)
+	bytes_push32f(Bytes self, float v)
 	{
 		[[maybe_unused]] size_t written_size = memory_stream_write(self, block_from(v));
 		assert(written_size == sizeof(v));
 	}
 
 	inline static void
-	bytes_push64f(Bytes& self, double v)
+	bytes_push64f(Bytes self, double v)
 	{
 		[[maybe_unused]] size_t written_size = memory_stream_write(self, block_from(v));
 		assert(written_size == sizeof(v));
 	}
 
 	inline static void
-	bytes_push_ptr(Bytes& self, void* ptr)
+	bytes_push_ptr(Bytes self, void* ptr)
 	{
 		uintptr_t v = (uintptr_t)ptr;
 		[[maybe_unused]] size_t written_size = memory_stream_write(self, block_from(v));
@@ -120,7 +120,7 @@ namespace mn
 	}
 
 	inline static void
-	bytes_push_block(Bytes& self, const Block& data)
+	bytes_push_block(Bytes self, Block data)
 	{
 		[[maybe_unused]] size_t written_size = memory_stream_write(self, data);
 		assert(data.size == written_size);
@@ -128,7 +128,7 @@ namespace mn
 
 
 	inline static uint8_t
-	bytes_pop8(Bytes& self)
+	bytes_pop8(Bytes self)
 	{
 		uint8_t res = 0;
 		[[maybe_unused]] size_t read_size = memory_stream_read(self, block_from(res));
@@ -137,7 +137,7 @@ namespace mn
 	}
 
 	inline static uint16_t
-	bytes_pop16(Bytes& self)
+	bytes_pop16(Bytes self)
 	{
 		uint16_t res = 0;
 		[[maybe_unused]] size_t read_size = memory_stream_read(self, block_from(res));
@@ -146,7 +146,7 @@ namespace mn
 	}
 
 	inline static uint32_t
-	bytes_pop32(Bytes& self)
+	bytes_pop32(Bytes self)
 	{
 		uint32_t res = 0;
 		[[maybe_unused]] size_t read_size = memory_stream_read(self, block_from(res));
@@ -155,7 +155,7 @@ namespace mn
 	}
 
 	inline static uint64_t
-	bytes_pop64(Bytes& self)
+	bytes_pop64(Bytes self)
 	{
 		uint64_t res = 0;
 		[[maybe_unused]] size_t read_size = memory_stream_read(self, block_from(res));
@@ -164,7 +164,7 @@ namespace mn
 	}
 
 	inline static float
-	bytes_pop32f(Bytes& self)
+	bytes_pop32f(Bytes self)
 	{
 		float res = 0;
 		[[maybe_unused]] size_t read_size = memory_stream_read(self, block_from(res));
@@ -173,7 +173,7 @@ namespace mn
 	}
 
 	inline static double
-	bytes_pop64f(Bytes& self)
+	bytes_pop64f(Bytes self)
 	{
 		double res = 0;
 		[[maybe_unused]] size_t read_size = memory_stream_read(self, block_from(res));
@@ -182,7 +182,7 @@ namespace mn
 	}
 
 	inline static void*
-	bytes_pop_ptr(Bytes& self)
+	bytes_pop_ptr(Bytes self)
 	{
 		void* res = 0;
 		[[maybe_unused]] size_t read_size = memory_stream_read(self, block_from(res));
@@ -191,15 +191,15 @@ namespace mn
 	}
 	
 	inline static void
-	bytes_pop_block(Bytes& self, const Block& data)
+	bytes_pop_block(Bytes self, Block data)
 	{
 		[[maybe_unused]] size_t read_size = memory_stream_read(self, data);
 		assert(read_size == data.size);
 	}
 
 	inline static Block
-	block_from(const Bytes& self)
+	block_from(Bytes self)
 	{
-		return block_from(self.str);
+		return block_from(self->str);
 	}
 }
