+++ mn/src/mn/Reader.cpp
@@ -1,6 +1,7 @@
 #include "mn/Reader.h"
 #include "mn/Stream.h"
 #include "mn/Memory_Stream.h"
+#include "mn/File.h"
 #include "mn/Pool.h"
 
 #include <assert.h>
@@ -20,7 +21,7 @@ namespace mn
 
 		Stdin_Reader_Wrapper()
 		{
-			self.stream = stream_stdin();
+			self.stream = file_stdin();
 			self.buffer = memory_stream_new();
 		}
 
@@ -80,7 +81,7 @@ namespace mn
 		assert(self->stream == nullptr);
 		memory_stream_clear(self->buffer);
 		memory_stream_write(self->buffer, Block { str.ptr, str.count });
-		str_null_terminate(self->buffer.str); //for strtoull bug since it need null terminated string
+		str_null_terminate(self->buffer->str); //for strtoull bug since it need null terminated string
 		memory_stream_cursor_to_start(self->buffer);
 		return self;
 	}
@@ -99,21 +100,21 @@ namespace mn
 	reader_peek(Reader self, size_t size)
 	{
 		//get the available data in the buffer
-		size_t available_size = self->buffer.str.count - self->buffer.cursor;
+		size_t available_size = self->buffer->str.count - self->buffer->cursor;
 
 		//if the user needs to peek on the already buffered data the return it as is
 		if(size == 0)
 			return memory_stream_block_ahead(self->buffer, available_size);
 
 		//save the old cursor
-		int64_t old_cursor = self->buffer.cursor;
+		int64_t old_cursor = self->buffer->cursor;
 		if(available_size < size)
 		{
 			size_t diff = size - available_size;
 			memory_stream_cursor_to_end(self->buffer);
 			if(self->stream)
 				available_size += memory_stream_pipe(self->buffer, self->stream, diff);
-			self->buffer.cursor = old_cursor;
+			self->buffer->cursor = old_cursor;
 		}
 		return memory_stream_block_ahead(self->buffer, available_size);
 	}
@@ -122,11 +123,11 @@ namespace mn
 	reader_skip(Reader self, size_t size)
 	{
 		//get the available data in the buffer
-		size_t available_size = self->buffer.str.count - self->buffer.cursor;
+		size_t available_size = self->buffer->str.count - self->buffer->cursor;
 
 		size_t result = available_size < size ? available_size : size;
 		memory_stream_cursor_move(self->buffer, result);
-		if(self->buffer.str.count - self->buffer.cursor == 0)
+		if(self->buffer->str.count - self->buffer->cursor == 0)
 			memory_stream_clear(self->buffer);
 		return result;
 	}
@@ -141,7 +142,7 @@ namespace mn
 		size_t request_size = data.size;
 		size_t read_size = 0;
 		//get the available data in the buffer
-		size_t available_size = self->buffer.str.count - self->buffer.cursor;
+		size_t available_size = self->buffer->str.count - self->buffer->cursor;
 		if(available_size > 0)
 		{
 			read_size += memory_stream_read(self->buffer, data);
